feat: add withActions and withVideo exports
