+++ src/views/homepage/components/hero/index.ts
@@ -1,0 +1,2 @@
+export { HeroWithActions } from './hero-with-actions'
+export { HeroWithVideo } from './hero-with-video'
